Fix using annonymous sumof

+++ src/test/kotlin/projectEuler/p16/PowerDigitSumTest.kt
@@ -1,11 +1,12 @@
 package projectEuler.p16
 
 import org.junit.jupiter.api.Test
+import java.math.BigInteger
 import kotlin.test.assertEquals
 
 class PowerDigitSumTest {
     @Test
     fun `2^15 results in 26`() {
-        assertEquals(26, PowerDigitSum.solve(2, 15))
+        assertEquals(BigInteger.valueOf(26), PowerDigitSum.solve(2, 15))
     }
 }
